:hammer_and_wrench: feat: extend Task interface to support packet size + count

+++ internal/model/Task.go
@@ -7,5 +7,5 @@ type Task interface {
 	Snapshot() interface{}
 	Reset()
 	Name() string
-	Query(flow []byte) uint32
+	Query(flow []byte) uint64
 }
